pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eda-agents"
version = "0.1.0"
description = "AI/LLM-assisted analog circuit design with SPICE-in-the-loop validation for IHP SG13G2"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
authors = [
    {name = "Mauricio-xx", email = "mauricio.xmonatanares@gmail.com"},
]
keywords = ["eda", "analog", "spice", "agents", "llm", "ihp-sg13g2", "circuit-design"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]
dependencies = [
    "numpy>=1.20",
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/Mauricio-xx/eda-agents"
Repository = "https://github.com/Mauricio-xx/eda-agents"
Issues = "https://github.com/Mauricio-xx/eda-agents/issues"

[project.optional-dependencies]
agents = [
    "openai>=1.0",
]
adk = [
    "google-adk>=1.0",
    "litellm>=1.50",
    "google-generativeai>=0.8",
]
coordination = [
    "context-teleport>=0.5.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "anyio>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/eda_agents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
anyio_mode = "auto"
markers = [
    "spice: requires ngspice + IHP SG13G2 PDK (deselect with -m 'not spice')",
]

[tool.ruff]
target-version = "py311"
line-length = 100
